% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_google_places.R
\name{add_google_places}
\alias{add_google_places}
\title{Append Google Places data to a dataframe}
\usage{
add_google_places(df, search_name, search_address, search_latitude,
  search_longitude, key = googleway:::get_api_key("places"),
  .keep_all = FALSE, ...)
}
\arguments{
\item{search_name}{Field containing the name of the location}

\item{search_address}{Field containing the address of the location}

\item{search_latitude}{Field containing the latitude of the location (for distance comparison)}

\item{search_longitude}{Field containing the longitude of the location (for distance comparison)}

\item{key}{A valid Google Developers Places API key.}

\item{.keep_all}{Toggle whether multiple results will be returned (if found)}

\item{...}{Other arguments passed to googleway::google_places}
}
\description{
This function takes a dataframe containing location information and
returns the same dataframe with additional enriched Google Places data.
If multiple results are found they will be compared/selected using
string similarity and geographic distance.
}
