% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_google_places.R
\name{add_tidy_google_places}
\alias{add_tidy_google_places}
\title{Append tidy google places to a dataframe}
\usage{
add_tidy_google_places(df, name, address, lat, lng, key)
}
\description{
Appends the best matching google places result to
a dataframe of lookup locations.
}
