% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_google_places.R
\name{get_tidy_google_place}
\alias{get_tidy_google_place}
\title{Get tidy Google Places information for a single location}
\usage{
get_tidy_google_place(search_name = NULL, search_address = NULL,
  search_latitude = NULL, search_longitude = NULL,
  key = googleway:::get_api_key("places"), .keep_all = FALSE, ...)
}
\arguments{
\item{search_name}{The name of the place to search}

\item{search_address}{The address (partial or full) of the place to search}

\item{search_latitude}{The latitude of the place to search (note - only used for
distance matching, not radial search)}

\item{search_longitude}{The latitude of the place to search (note - only used for
distance matching, not radial search)}

\item{key}{A valid Google Developers Places API key.}

\item{.keep_all}{Toggle whether multiple results will be returned (if found)}

\item{...}{Other arguments passed to googleway::google_places}
}
\description{
This function calls the Google Places API using the
googleway::google_places function. It will return a single,
tidy dataframe instead of a json object. Result selection
is done using string and geographic distances.
}
