% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_google_places.R
\name{tidy_google_places}
\alias{tidy_google_places}
\title{Tidy Google Places}
\usage{
tidy_google_places(search_name = NULL, search_address = NULL,
  search_lat = NULL, search_lng = NULL, keep_top = TRUE,
  key = NULL, ...)
}
\arguments{
\item{search_name}{The name of the place to search}

\item{search_address}{The address (partial or full) of the place to search}

\item{search_lat}{The latitude of the place to search (note - only used for
distance matching, not radial search - use location)}

\item{search_lng}{The latitude of the place to search (note - only used for
distance matching, not radial search - use location)}

\item{keep_top}{Should only the best matching result be returned?}

\item{key}{\code{string} A valid Google Developers Places API key.}

\item{...}{Other arguments passed to googleway}
}
\description{
This function calls the Google Places API using the
googleway::google_places function. It will return a single,
tidy dataframe instead of a json object. Result selection
is done using string distance similarity.
}
